Calculates weights for terms and documents. Assumes 1000 docs in collection, since can't currently calculate this.
git-svn-id: file:///home/xapian-svn/xapian/trunk@60 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/common/omtypes.h
@@ -8,7 +8,7 @@ typedef unsigned int docid;
 
 typedef docid doccount;
 
-typedef unsigned int weight;
+typedef double weight;
 
 #ifdef __cplusplus
 typedef string termname;
